Continue editing from ByeCommand
