+++ src/main/java/command/MarkCommand.java
@@ -1,22 +1,23 @@
-package Command;
+package command;
 
-import Task.TaskList;
-import Main.UI;
-import Main.Storage;
+import task.TaskList;
+import main.UI;
+import main.Storage;
 
-import Exception.DukeException;
+import exception.DukeException;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
 
     int index;
-    public UnmarkCommand(int index) {
+
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, UI ui, Storage storage) throws DukeException {
         try {
-            taskList.unmark(this.index);
+            taskList.mark(this.index);
             storage.saveList(taskList);
         } catch (RuntimeException e) {
             throw new DukeException(String.format("Given index is out of range. Index range should be between" +
