+++ src/main/java/command/UnmarkCommand.java
@@ -1,23 +1,22 @@
-package Command;
+package command;
 
-import Task.TaskList;
-import Main.UI;
-import Main.Storage;
+import task.TaskList;
+import main.UI;
+import main.Storage;
 
-import Exception.DukeException;
+import exception.DukeException;
 
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
 
     int index;
-
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, UI ui, Storage storage) throws DukeException {
         try {
-            taskList.mark(this.index);
+            taskList.unmark(this.index);
             storage.saveList(taskList);
         } catch (RuntimeException e) {
             throw new DukeException(String.format("Given index is out of range. Index range should be between" +
