+++ src/main/java/exception/DukeException.java
@@ -1,4 +1,4 @@
-package Exception;
+package exception;
 
 public class DukeException extends Exception {
     public DukeException(String message) {
