+++ src/main/java/main/UI.java
@@ -1,6 +1,6 @@
-package Main;
+package main;
 
-import Task.Task;
+import task.Task;
 
 import java.util.ArrayList;
