+++ src/main/java/task/TaskList.java
@@ -1,6 +1,6 @@
-package Task;
+package task;
 
-import Exception.DukeException;
+import exception.DukeException;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
