+++ src/test/java/task/TaskListTest.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 
 import org.junit.jupiter.api.Test;
